Break the API out into a separate app
