+++ bongo/apps/bongo/management/commands/gentags.py
@@ -1,4 +1,4 @@
-from django.core.management.base import BaseCommand
+from django.bongo.management.base import BaseCommand
 from bongo.apps.bongo.models import Post, Text
 from bongo.apps.bongo.helpers import tagify
