+++ bongo/apps/bongo/management/commands/nltk-init.py
@@ -1,4 +1,4 @@
-from django.core.management.base import BaseCommand
+from django.bongo.management.base import BaseCommand
 import nltk
 
 class Command(BaseCommand):
